+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Add `Chunking Master` quest
+
 ## [0.1.0] - 2026-02-25
 
 ### Changed
@@ -55,10 +59,17 @@
   from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 
 [Unreleased]: https://github.com/ythirion/wor/compare/0.1.0...HEAD
+
 [0.1.0]: https://github.com/ythirion/wor/compare/0.0.7...0.1.0
+
 [0.0.7]: https://github.com/ythirion/wor/compare/0.0.6...0.0.7
+
 [0.0.6]: https://github.com/ythirion/wor/compare/0.0.5...0.0.6
+
 [0.0.5]: https://github.com/ythirion/wor/compare/0.0.4...0.0.5
+
 [0.0.4]: https://github.com/ythirion/wor/compare/0.0.3...0.0.4
+
 [0.0.3]: https://github.com/ythirion/wor/compare/0.0.1...0.0.3
+
 [0.0.1]: https://github.com/ythirion/wor/commits/0.0.1